refactor; split QueryBuilder into QueryBuilder and QueryExecutor
