+++ cognite/pygen/_query/interface.py
@@ -13,8 +13,9 @@ from cognite.client.utils.useful_types import SequenceNotStr
 
 from cognite.pygen._version import __version__
 
-from .builder import QueryBuilder, chunker
+from .builder import QueryBuilder
 from .constants import AGGREGATION_LIMIT, IN_FILTER_CHUNK_SIZE, SEARCH_LIMIT, SelectedProperties
+from .executer import chunker
 from .processing import QueryUnpacker
 from .step import QueryBuildStepFactory
 
@@ -242,7 +243,8 @@ class QueryExecutor:
         builder.append(factory.root(filter, limit=limit))
         for connection_id, connection in factory.connection_properties.items():
             builder.extend(factory.from_connection(connection_id, connection, reverse_views))
-        _ = builder.execute_query(self._client, remove_not_connected=False)
+        executor = builder.build()
+        _ = executor.execute_query(self._client, remove_not_connected=False)
         return QueryUnpacker(
             builder, edges=self._unpack_edges, as_data_record=False, edge_type_key="type", node_type_key="type"
         ).unpack()
